ex03 Ok

+++ P1/ex03/load_image.py
@@ -6,7 +6,7 @@ import numpy as np
 def ft_load(path: str) -> np.array:
     try:
         img = Image.open(path)
-        array_img = np.array(img)
+        array_img = np.array(img, dtype=np.uint8)
         print(f"The shape of image is: {array_img.shape}")
         print(array_img)
         img_gray = img.convert('L')
